feat(docs): Add TOC to supported hardware page.
